Consistency/Flow (#342)
Fixes an error with the chronological appearance of code fragments to keep the code consistent and improve the overall flow. The `err = console.error` fragment doesn't appear until the `Handling errors` section of the lesson, so it is removed in previous sections (where it is unnecessary, anyway).

I figure it's either this, or the `Handling errors` section should somehow come _before_ the `Accepting a callback function` section.